+++ packages/expo-ui/src/jetpack-compose/index.ts
@@ -17,6 +17,7 @@ export * from './ExposedDropdownMenuBox';
 export * from './Divider';
 export * from './Host';
 export * from './LazyColumn';
+export { type LazyItemsProps } from './LazyItems';
 export * from './LazyRow';
 export * from './ListItem';
 export * from './RNHostView';